DATN1/warehouse/flask-template/requirements.txt
flask
Pillow
opencv-python
flask-socketio 
eventlet
flask-sqlalchemy 
flask-wtf
werkzeug